feat(GAME-082): rename scenario-010 → tutorial-003, expand to 6×6, all terrain features + foothill rendering
